requirements.txt
python-telegram-bot==20.0
APScheduler==3.9.1
python-dotenv==0.19.2
schedule==1.2.0
